reorganize files for better app architecture
